move go extraction rules behind the indexer language registry
S3-GO (ARCH-012), one language rollback unit. Go's tree-sitter rule file,
compiled-rule cache, selector-call callsite marker, manual MEMBER-edge
collector, and receiver-call resolution engine move out of the indexer god file
into `codestory-indexer/src/languages/go.rs`, behind the `LanguageExtraction`
registry in `languages/mod.rs`. Every language-keyed dispatch that used to
spell "go" now reads the registry first and keeps its residual `match` for the
languages that have not moved; the last sibling package deletes those arms.

The move is mechanical: 31 functions left `lib.rs` and the same 31 landed in
`languages/go.rs`. Diffing the moved bodies against the deleted lines, ignoring
blank lines, is empty once three declared renames are applied
(`collect_go_member_edges` -> `member_edge_specs`,
`collect_go_receiver_call_edges` -> `receiver_call_specs`, `go_selector_call`
-> `selector_call`) plus `pub(crate)` on the two registry entry points.
`insert_unique_import_binding` moved with them because Go's import scanner was
its only caller; `descendant_by_field_name`, `first_descendant_with_kind`,
`normalized_receiver_variable`, and `node_is_same_or_ancestor` stayed because
other languages use them.

Go needed one field the Kotlin row did not: `member_edge_specs`. Go is the
first migrated language with a manual MEMBER-edge collector, so
`lib.rs::language_member_specs` had no registry lookup at all. Without the
field, the "go" arm would have had to stay in that match and Go would still
appear twice. Kotlin's row is `None`, which is exactly what the residual
`_ => Vec::new()` arm gave it, so the addition is behaviour-neutral.

Proof: `tests/language_extraction_snapshot.rs` gains a Go row pinning the full
rendered projection of both Go fixtures - the fidelity_lab file through the
store pipeline (resolution included) and the tictactoe file through raw
`index_file`. Both goldens were captured before the production edit and are
byte-identical after (sha256 746d7345.. / 8102ed77..). Falsified by mutation:
setting `receiver_call_specs: None`, `qualified_name_delimiter: "::"`, or
`member_edge_specs: None` each turns both snapshots red, and the delimiter
change leaves `fidelity_regression` and `tictactoe_language_coverage` green -
no pre-existing test catches it.

Declared behavior changes:
- None to any indexing, resolution, or rendering output. Go fidelity_regression
  and tictactoe_language_coverage projections are proven byte-identical before
  and after, and packet_search_eval is unmoved.
- `LanguageExtraction` gains `member_edge_specs`; `lib.rs::language_member_specs`
  becomes registry-first. Kotlin's row is `None`, matching the residual arm it
  used to fall through to.
- `LanguageRuleset::Go`'s arm in `compiled_rules` is now unreachable and returns
  a typed error instead of compiled rules; the registry answers first.
  Fail-closed by construction, matching the Kotlin arm.
- `codestory_contracts::language_support::LANGUAGE_COMMENT_PROFILES` gains a
  `go` row and `codestory-cli`'s local comment roster drops its `go` entry. The
  marker is `//` on both sides, so rendering is unchanged; the CLI roster's
  deliberate disagreements with the indexer route scanner (cpp/swift/tsx/jsx
  present, c absent) are untouched.
- `LANGUAGE_SUPPORT_PROFILES` is unchanged, so extension routing, claim tiers,
  `supported_extensions`, `is_structural_source_path`, and every existing
  language claim are identical.
- `SemanticResolverKind::Go` stays in `semantic::dedicated_semantic_resolver`
  because Go has a dedicated resolver type and those types are private to that
  module; the row records the choice with
  `uses_generic_semantic_resolver: false`.
- No version surface touched; everything stays 0.16.3.

Closes #1685
Refs #1243
Refs #1179
