+++ crates/codestory-indexer/src/resolution/mod.rs
@@ -1393,7 +1393,7 @@ fn is_go_selector_call_placeholder(edge_kind: EdgeKind, callsite_identity: Optio
     callsite_identity.is_some_and(|identity| {
         identity
             .split('|')
-            .any(|part| part == crate::GO_SELECTOR_CALLSITE_MARKER)
+            .any(|part| part == crate::languages::go::MEMBER_CALLSITE_MARKER)
     })
 }
